# docs/contracts/local-api.openapi.yml
openapi: 3.0.3
info:
  title: Sistema MNSS Local API
  version: 0.0.1
  description: |
    API para operação local da Padaria Nova Aliança.
    
    ### WebSocket Events
    O sistema publica eventos em tempo real para sincronização de telas:
    
    - `/topic/kds/tickets`: Enviado quando um ticket é criado ou atualizado. Payload: `KdsTicketResponse`.
    - `/topic/kds/tickets/{sector}`: Mesmo evento acima, mas filtrado por setor (ex: `COZINHA`, `FORNO`).
    - `/topic/orders/ready`: Enviado quando um pedido completo está pronto para entrega. Payload: `UUID` (ID do pedido).
paths:
...
  /api/users:
    get:
      summary: Lista usuarios internos.
      operationId: listInternalUsers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Usuarios cadastrados.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
        '403':
          $ref: '#/components/responses/BusinessError'
        default:
          $ref: '#/components/responses/UnexpectedError'
    post:
      summary: Cria usuario interno.
      operationId: createInternalUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '200':
          description: Usuario criado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          $ref: '#/components/responses/BusinessError'
        '409':
          $ref: '#/components/responses/BusinessError'
        default:
          $ref: '#/components/responses/UnexpectedError'
...
  /api/stock-movements:
    get:
      summary: Lista movimentacoes de estoque.
      operationId: listStockMovements
      security:
        - bearerAuth: []
      parameters:
        - name: productId
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Movimentacoes registradas.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StockMovementResponse'
        '403':
          $ref: '#/components/responses/BusinessError'
        default:
          $ref: '#/components/responses/UnexpectedError'
    post:
      summary: Registra movimentacao de estoque local.
      operationId: createStockMovement
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StockMovementPayload'
      responses:
        '201':
          description: Movimentacao registrada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockMovementResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          $ref: '#/components/responses/BusinessError'
        '404':
          $ref: '#/components/responses/BusinessError'
        default:
          $ref: '#/components/responses/UnexpectedError'
  /api/stock-movements/balances:
    get:
      summary: Lista saldo total de estoque por produto.
      operationId: listStockBalances
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Saldos calculados por produto.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StockBalanceResponse'
        '403':
          $ref: '#/components/responses/BusinessError'
        default:
          $ref: '#/components/responses/UnexpectedError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    ValidationError:
      description: Requisicao invalida por erro de validacao.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    BusinessError:
      description: Regra de negocio impediu a operacao.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    UnexpectedError:
      description: Erro inesperado sem vazamento de detalhes internos.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
    ApiError:
      type: object
      required:
        - code
        - message
        - status
        - path
        - timestamp
        - validationErrors
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Campos invalidos na requisicao.
        status:
          type: integer
          example: 400
        path:
          type: string
          example: /api/ping
        timestamp:
          type: string
          format: date-time
        validationErrors:
          type: array
          items:
            $ref: '#/components/schemas/ApiFieldError'
    ApiFieldError:
      type: object
      required:
        - field
        - message
      properties:
        field:
          type: string
          example: name
        message:
          type: string
          example: must not be blank
...
    CreateUserRequest:
      type: object
      required:
        - name
        - username
        - password
        - roles
      properties:
        name:
          type: string
          example: Operador Caixa
        email:
          type: string
          example: caixa@novaalianca.local
        username:
          type: string
          example: caixa01
        password:
          type: string
          format: password
        active:
          type: boolean
          example: true
        roles:
          type: array
          items:
            $ref: '#/components/schemas/RoleName'
          example: [CAIXA]
    UserResponse:
      type: object
      required:
        - id
        - name
        - username
        - active
        - roles
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
        username:
          type: string
        active:
          type: boolean
        roles:
          type: array
          items:
            $ref: '#/components/schemas/RoleName'
...
    StockMovementPayload:
      type: object
      required:
        - productId
        - type
        - quantity
      properties:
        productId:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/StockMovementType'
        quantity:
          type: number
          format: decimal
          minimum: 0.001
        reason:
          type: string
        orderId:
          type: string
          format: uuid
    StockMovementResponse:
      allOf:
        - $ref: '#/components/schemas/StockMovementPayload'
        - type: object
          required:
            - id
            - productName
            - createdBy
            - createdAt
            - updatedAt
          properties:
            id:
              type: string
              format: uuid
            productName:
              type: string
            createdBy:
              type: string
              format: uuid
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
    StockBalanceResponse:
      type: object
      required:
        - productId
        - productName
        - quantity
      properties:
        productId:
          type: string
          format: uuid
        productName:
          type: string
        quantity:
          type: number
          format: decimal
    StockMovementType:
      type: string
      enum:
        - IN
        - OUT
        - ADJUSTMENT
        - SALE
        - LOSS
        - RETURN
    RoleName:
      type: string
      enum:
        - ADMIN
        - GERENTE
        - CAIXA
        - ATENDENTE
        - COZINHA
        - ENTREGADOR
        - CONSULTA
...